fix(csp,consent): add dangerous unsafe-inline csp rule for script-src
TH-1461.

Cookie consent modal is unusable without allowing inline-scripting.
Since we don't know yet how to explicitly fix it, we need to allow all
"unsafe-inline" scripting until we find a solution. It's dangerous in
context of CSP, but quite safe in context of events-monorepo app, since
there is no authentication or personal data of any kind. Also the
content of it is curated.