fix(testcafe): use --disable-features=LocalNetworkAccessChecks
Chrome 140→142 broke testcafe browser tests, see
https://github.com/DevExpress/testcafe/issues/8446

Testcafe has not released a bugfix for this issue as of 2025-12-02, but
the comments in the above issue pointed to using
`--disable-features=LocalNetworkAccessChecks` with chrome as a working
workaround.

This PR only adds `--disable-features=LocalNetworkAccessChecks` to all
testcafe invocations, and removes wsl2win testcafe running as
unnecessary.

refs LIIKUNTA-770